codebase: Move gitbot command overrides into their own module
