+++ lib/structs/__init__.py
@@ -9,3 +9,4 @@ from .caches.self_hashing_cache import SelfHashingCache
 from .caches.base_cache import BaseCache
 from .db.user_collection import UserCollection
 from .discord.gitbot_embed import *
+from .discord.commands import *
